pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "watermark"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Peter Egger", email = "peter.pfaffeneder@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pillow"
]

[project.urls]
Documentation = "https://github.com/Peter Egger/watermark#readme"
Issues = "https://github.com/Peter Egger/watermark/issues"
Source = "https://github.com/Peter Egger/watermark"

[tool.hatch.version]
path = "src/watermark/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/watermark tests}"

[tool.coverage.run]
source_pkgs = ["watermark", "tests"]
branch = true
parallel = true
omit = [
  "src/watermark/__about__.py",
]

[tool.coverage.paths]
watermark = ["src/watermark", "*/watermark/src/watermark"]
tests = ["tests", "*/watermark/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
